Reordered projects page into components, and moved related components into projects/components
