+++ src/app/projects/page.tsx
@@ -1,9 +1,8 @@
 'use client';
 
 import { useState, useEffect, useCallback, useRef } from 'react';
-import Image from 'next/image';
 import { useSession } from 'next-auth/react';
-import ProjectUploadForm from '@/components/ProjectUploadForm';
+import ProjectUploadForm from './components/ProjectUploadForm';
 import {
   DndContext,
   closestCenter,
@@ -20,252 +19,16 @@ import {
   arrayMove,
   SortableContext,
   rectSortingStrategy,
-  useSortable,
 } from '@dnd-kit/sortable';
-import { CSS } from '@dnd-kit/utilities';
 import type { ProjectApiResponse } from '@/types/project';
 
-/** History state for modal stack: each entry is one modal level (gallery or photo). */
-type ModalHistoryState =
-  | { modal: 'gallery'; projectId: string }
-  | { modal: 'photo'; projectId: string; photoIndex: number }
-  | null;
-
-const PAGE_SIZE = 6;
-const TAG_NONE = '__none__';
-
-function formatProjectDate(createdAt: string, dateIsMonthOnly?: boolean | null): string {
-  const d = new Date(createdAt);
-  const month = d.toLocaleString('default', { month: 'long' });
-  const year = d.getFullYear();
-  if (dateIsMonthOnly) {
-    return `${month} ${year}`;
-  }
-  const day = d.getDate();
-  return `${month} ${day}, ${year}`;
-}
-
-function ProjectCard({
-  project,
-  priority = false,
-  isAuthenticated = false,
-  onEdit,
-  onDelete,
-  onPhotoClick,
-  onGalleryClick,
-  deletingId,
-}: {
-  project: ProjectApiResponse;
-  priority?: boolean;
-  isAuthenticated?: boolean;
-  onEdit?: (project: ProjectApiResponse) => void;
-  onDelete?: (id: string) => void;
-  onPhotoClick?: (project: ProjectApiResponse) => void;
-  onGalleryClick?: (project: ProjectApiResponse) => void;
-  deletingId?: string | null;
-}) {
-  const handleImageClick = () => {
-    if (project.imageUrl && onPhotoClick) onPhotoClick(project);
-  };
-  const handleOverlayClick = (e: React.MouseEvent | React.TouchEvent) => {
-    e.preventDefault();
-    e.stopPropagation();
-    if (onGalleryClick) onGalleryClick(project);
-  };
-
-  return (
-    <div className="bg-card border border-border rounded-xl shadow-md overflow-hidden hover:shadow-lg transition-shadow h-full flex flex-col">
-      <div className="relative h-72 bg-muted shrink-0">
-        {project.imageUrl ? (
-          <button
-            type="button"
-            onClick={handleImageClick}
-            className="absolute inset-0 w-full h-full cursor-zoom-in focus:outline-none focus-visible:ring-2 focus-visible:ring-primary focus-visible:ring-offset-2 focus-visible:ring-offset-background rounded-t-xl touch-manipulation"
-            aria-label={`View full size: ${project.title}`}
-          >
-            <Image
-              src={project.imageUrl}
-              alt={project.title}
-              fill
-              sizes="(max-width: 768px) 100vw, (max-width: 1024px) 50vw, 33vw"
-              className="object-cover pointer-events-none"
-              priority={priority}
-            />
-          </button>
-        ) : (
-          <div className="absolute inset-0 flex items-center justify-center text-muted-foreground/50 text-sm">
-            No image
-          </div>
-        )}
-        {project.featured && (
-          <span className="absolute top-2 right-2 bg-accent text-accent-foreground px-2 py-1 rounded text-xs font-medium shadow">
-            Featured
-          </span>
-        )}
-        {/* Overlay: title, tags, date — click/tap opens gallery modal */}
-        <button
-          type="button"
-          onClick={handleOverlayClick}
-          className="absolute inset-x-0 bottom-0 w-full text-left bg-gradient-to-t from-black/85 via-black/60 to-transparent pt-12 px-3 pb-3 cursor-pointer focus:outline-none focus-visible:ring-2 focus-visible:ring-primary focus-visible:ring-inset touch-manipulation"
-          aria-label={`Open gallery: ${project.title}`}
-        >
-          <h4 className="font-bold text-lg text-white drop-shadow-sm">{project.title}</h4>
-          <div className="min-h-[1.5rem] flex flex-wrap gap-1.5 mt-1.5 mb-1">
-            {project.tags?.map((tag) => (
-              <span
-                key={tag}
-                className="inline-block px-2 py-0.5 rounded bg-white/25 text-white text-xs backdrop-blur-sm"
-              >
-                {tag}
-              </span>
-            ))}
-          </div>
-          <p className="text-xs text-white/90">
-            {formatProjectDate(project.createdAt, project.dateIsMonthOnly)}
-          </p>
-        </button>
-      </div>
-      {project.description && (
-        <div className="p-4 pt-2">
-          <p className="text-sm text-foreground/80 line-clamp-3">{project.description}</p>
-        </div>
-      )}
-      {isAuthenticated && onEdit && onDelete && (
-        <div className="flex gap-2 p-4 pt-2 mt-auto border-t border-border">
-          <button
-            type="button"
-            onClick={() => onEdit(project)}
-            className="flex-1 bg-primary text-primary-foreground py-2 px-4 rounded-lg font-medium hover:bg-accent transition-colors text-sm"
-          >
-            Edit
-          </button>
-          <button
-            type="button"
-            onClick={() => onDelete(project.id)}
-            disabled={deletingId === project.id}
-            className="flex-1 bg-destructive text-destructive-foreground py-2 px-4 rounded-lg font-medium hover:bg-destructive/90 transition-colors disabled:opacity-50 disabled:cursor-not-allowed text-sm"
-          >
-            {deletingId === project.id ? 'Deleting…' : 'Delete'}
-          </button>
-        </div>
-      )}
-    </div>
-  );
-}
-
-function SortableProjectCard({
-  project,
-  dateKeyOfDragged,
-  isAuthenticated = false,
-  onEdit,
-  onDelete,
-  deletingId,
-}: {
-  project: ProjectApiResponse;
-  dateKeyOfDragged: string | null;
-  isAuthenticated?: boolean;
-  onEdit?: (project: ProjectApiResponse) => void;
-  onDelete?: (id: string) => void;
-  deletingId?: string | null;
-}) {
-  const dateKey = project.createdAt?.slice(0, 10) ?? '';
-  const isSameDateGroup = dateKeyOfDragged == null || dateKey === dateKeyOfDragged;
-
-  const {
-    attributes,
-    listeners,
-    setNodeRef,
-    transform,
-    transition,
-    isDragging,
-  } = useSortable({
-    id: project.id,
-    data: { dateKey },
-  });
-
-  const style = {
-    transform: CSS.Transform.toString(transform),
-    transition,
-  };
-
-  return (
-    <div
-      ref={setNodeRef}
-      style={style}
-      {...(isSameDateGroup ? { ...attributes, ...listeners } : {})}
-      className={`bg-card border border-border rounded-xl overflow-hidden flex flex-col h-full touch-manipulation select-none ${
-        !isSameDateGroup
-          ? 'opacity-70 pointer-events-none cursor-default ring-0 shadow-none'
-          : isDragging
-            ? 'opacity-90 z-10 cursor-grabbing ring-2 ring-primary ring-offset-2 ring-offset-background shadow-[0_0_0_2px_var(--primary),0_0_24px_color-mix(in_srgb,var(--primary)_50%,transparent)]'
-            : 'cursor-grab active:cursor-grabbing ring-2 ring-primary/50 ring-offset-2 ring-offset-background shadow-[0_0_0_1px_color-mix(in_srgb,var(--primary)_30%,transparent),0_0_16px_color-mix(in_srgb,var(--primary)_20%,transparent)] hover:ring-primary/70 hover:shadow-[0_0_0_1px_color-mix(in_srgb,var(--primary)_40%,transparent),0_0_20px_color-mix(in_srgb,var(--primary)_30%,transparent)] transition-all'
-      }`}
-      aria-label={isSameDateGroup ? `Drag to reorder: ${project.title}` : `${project.title} (different date)`}
-    >
-      <div className="relative h-72 bg-muted shrink-0">
-        {project.imageUrl ? (
-          <Image
-            src={project.imageUrl}
-            alt={project.title}
-            fill
-            sizes="(max-width: 768px) 100vw, (max-width: 1024px) 50vw, 33vw"
-            className="object-cover pointer-events-none"
-          />
-        ) : (
-          <div className="absolute inset-0 flex items-center justify-center text-muted-foreground/50 text-sm">
-            No image
-          </div>
-        )}
-        {project.featured && (
-          <span className="absolute top-2 right-2 bg-accent text-accent-foreground px-2 py-1 rounded text-xs font-medium shadow">
-            Featured
-          </span>
-        )}
-        {/* Overlay: title, tags, date — pointer-events-none so taps pass through to open photo modal */}
-        <div className="absolute inset-x-0 bottom-0 bg-gradient-to-t from-black/85 via-black/60 to-transparent pt-12 px-3 pb-3 pointer-events-none">
-          <h4 className="font-bold text-lg text-white drop-shadow-sm">{project.title}</h4>
-          <div className="min-h-[1.5rem] flex flex-wrap gap-1.5 mt-1.5 mb-1">
-            {project.tags?.map((tag) => (
-              <span
-                key={tag}
-                className="inline-block px-2 py-0.5 rounded bg-white/25 text-white text-xs backdrop-blur-sm"
-              >
-                {tag}
-              </span>
-            ))}
-          </div>
-          <p className="text-xs text-white/90">
-            {formatProjectDate(project.createdAt, project.dateIsMonthOnly)}
-          </p>
-        </div>
-      </div>
-      {project.description && (
-        <div className="p-4 pt-2">
-          <p className="text-sm text-foreground/80 line-clamp-3">{project.description}</p>
-        </div>
-      )}
-      {isAuthenticated && onEdit && onDelete && (
-        <div className="flex gap-2 p-4 pt-2 mt-auto border-t border-border">
-          <button
-            type="button"
-            onClick={() => onEdit(project)}
-            className="flex-1 bg-primary text-primary-foreground py-2 px-4 rounded-lg font-medium hover:bg-accent transition-colors text-sm"
-          >
-            Edit
-          </button>
-          <button
-            type="button"
-            onClick={() => onDelete(project.id)}
-            disabled={deletingId === project.id}
-            className="flex-1 bg-destructive text-destructive-foreground py-2 px-4 rounded-lg font-medium hover:bg-destructive/90 transition-colors disabled:opacity-50 disabled:cursor-not-allowed text-sm"
-          >
-            {deletingId === project.id ? 'Deleting…' : 'Delete'}
-          </button>
-        </div>
-      )}
-    </div>
-  );
-}
+import { PAGE_SIZE, TAG_NONE, type ModalHistoryState } from './constants';
+import ProjectCard from './components/ProjectCard';
+import SortableProjectCard from './components/SortableProjectCard';
+import ProjectsFilters from './components/ProjectsFilters';
+import ProjectsReorderBar from './components/ProjectsReorderBar';
+import GalleryModal from './components/GalleryModal';
+import PhotoModal from './components/PhotoModal';
 
 export default function ProjectsPage() {
   const { status } = useSession();
@@ -289,8 +52,6 @@ export default function ProjectsPage() {
   const [photoModalIndex, setPhotoModalIndex] = useState(0);
   const [photoModalDragOffset, setPhotoModalDragOffset] = useState(0);
   const [photoModalIsDragging, setPhotoModalIsDragging] = useState(false);
-  const touchStartX = useRef<number | null>(null);
-  const touchStartOffset = useRef(0);
   const carouselContainerRef = useRef<HTMLDivElement | null>(null);
   const [galleryModalProject, setGalleryModalProject] = useState<ProjectApiResponse | null>(null);
   const projectsRef = useRef<ProjectApiResponse[]>([]);
@@ -379,7 +140,6 @@ export default function ProjectsPage() {
     fetchYears();
   }, [fetchTagNames, fetchYears]);
 
-  // Lock body scroll when photo or gallery modal is open (mobile-friendly)
   useEffect(() => {
     if (photoModalProject || galleryModalProject) {
       const prev = document.body.style.overflow;
@@ -390,7 +150,6 @@ export default function ProjectsPage() {
     }
   }, [photoModalProject, galleryModalProject]);
 
-  // Sync modal state from history when user presses back/forward (popstate)
   useEffect(() => {
     const onPopState = (e: PopStateEvent) => {
       const state = e.state as ModalHistoryState | null | undefined;
@@ -447,7 +206,6 @@ export default function ProjectsPage() {
     setGalleryModalProject(null);
   }, []);
 
-  // Close gallery on Escape
   useEffect(() => {
     if (!galleryModalProject) return;
     const onKeyDown = (e: KeyboardEvent) => {
@@ -457,7 +215,6 @@ export default function ProjectsPage() {
     return () => window.removeEventListener('keydown', onKeyDown);
   }, [galleryModalProject, closeGalleryModal]);
 
-  // Close photo modal on Escape; Arrow keys for carousel
   useEffect(() => {
     if (!photoModalProject) return;
     const imageUrls =
@@ -472,7 +229,6 @@ export default function ProjectsPage() {
     return () => window.removeEventListener('keydown', onKeyDown);
   }, [photoModalProject, closePhotoModal]);
 
-  // Mouse: small movement to start drag. Touch: long-press (250ms) so scroll works until then.
   const sensors = useSensors(
     useSensor(MouseSensor, { activationConstraint: { distance: 8 } }),
     useSensor(TouchSensor, { activationConstraint: { delay: 250, tolerance: 5 } }),
@@ -589,12 +345,25 @@ export default function ProjectsPage() {
     [refreshProjects],
   );
 
+  const handleCancelReorder = useCallback(async () => {
+    const count = projects.length;
+    setIsReorderMode(false);
+    setSelectedTag(null);
+    await fetchProjects({ limit: Math.max(PAGE_SIZE, count) });
+  }, [projects.length, fetchProjects]);
+
+  const photoModalImageUrls =
+    photoModalProject?.images?.length
+      ? photoModalProject.images.map((i) => i.imageUrl)
+      : photoModalProject?.imageUrl
+        ? [photoModalProject.imageUrl]
+        : [];
+
   return (
     <div className="pt-16 min-h-screen">
       <div className="max-w-7xl mx-auto px-4 py-12">
         <h1 className="text-4xl font-bold mb-4 text-foreground">Projects</h1>
 
-        {/* Project upload form (authenticated only) */}
         {isAuthenticated && (
           <div className="mb-8 sm:mb-12">
             {editingProject && (
@@ -619,105 +388,24 @@ export default function ProjectsPage() {
           </div>
         )}
 
-        {/* Tag + Year filters + Edit order (when authenticated) */}
-        {!isReorderMode && (
-          <div className="flex flex-wrap items-center gap-3 mb-8">
-            <div className="flex flex-wrap items-center gap-2">
-              <span className="text-sm text-muted-foreground">Tag:</span>
-              <button
-                onClick={() => setSelectedTag(null)}
-                className={`px-3 py-1.5 rounded-lg text-sm font-medium transition-colors ${
-                  selectedTag === null
-                    ? 'bg-primary text-primary-foreground'
-                    : 'bg-muted text-foreground hover:bg-muted/80'
-                }`}
-              >
-                All
-              </button>
-              <button
-                onClick={() => setSelectedTag(TAG_NONE)}
-                className={`px-3 py-1.5 rounded-lg text-sm font-medium transition-colors ${
-                  selectedTag === TAG_NONE
-                    ? 'bg-primary text-primary-foreground'
-                    : 'bg-muted text-foreground hover:bg-muted/80'
-                }`}
-              >
-                None
-              </button>
-              {tagNames.map((name) => (
-                <button
-                  key={name}
-                  onClick={() => setSelectedTag(name)}
-                  className={`px-3 py-1.5 rounded-lg text-sm font-medium transition-colors ${
-                    selectedTag === name
-                      ? 'bg-primary text-primary-foreground'
-                      : 'bg-muted text-foreground hover:bg-muted/80'
-                  }`}
-                >
-                  {name}
-                </button>
-              ))}
-            </div>
-            <div className="flex flex-wrap items-center gap-2">
-              <span className="text-sm text-muted-foreground">Year:</span>
-              <select
-                aria-label="Filter by year"
-                value={selectedYear ?? ''}
-                onChange={(e) =>
-                  setSelectedYear(e.target.value ? parseInt(e.target.value, 10) : null)
-                }
-                className="px-3 py-1.5 rounded-lg text-sm font-medium border border-border bg-card text-foreground focus:outline-none focus:ring-2 focus:ring-ring"
-              >
-                <option value="">All time</option>
-                {years.map((y) => (
-                  <option key={y} value={y}>
-                    {y}
-                  </option>
-                ))}
-              </select>
-            </div>
-          </div>
-        )}
-        {isReorderMode && (
-          <p className="text-muted-foreground mb-4">
-            {selectedYear !== null
-              ? `Showing all projects from ${selectedYear} in true order. Drag cards to reorder.`
-              : 'Showing all projects in true order. Drag cards to reorder.'}
-          </p>
-        )}
+        <ProjectsFilters
+          tagNames={tagNames}
+          years={years}
+          selectedTag={selectedTag}
+          selectedYear={selectedYear}
+          isReorderMode={isReorderMode}
+          onTagChange={setSelectedTag}
+          onYearChange={setSelectedYear}
+        />
+
         {isAuthenticated && !loading && (projects.length > 0 || isReorderMode) && (
-          <div className="flex flex-wrap items-center justify-end gap-2 mb-8">
-            {isReorderMode ? (
-              <>
-                <button
-                  onClick={async () => {
-                    const count = projects.length;
-                    setIsReorderMode(false);
-                    setSelectedTag(null);
-                    await fetchProjects({ limit: Math.max(PAGE_SIZE, count) });
-                  }}
-                  disabled={isSavingOrder}
-                  className="px-4 py-2 rounded-lg border border-border bg-card text-foreground font-medium hover:bg-muted transition-colors disabled:opacity-50"
-                >
-                  Cancel
-                </button>
-                <button
-                  onClick={handleSaveOrder}
-                  disabled={isSavingOrder}
-                  className="px-4 py-2 rounded-lg bg-primary text-primary-foreground font-medium hover:bg-accent transition-colors disabled:opacity-50 disabled:cursor-not-allowed"
-                >
-                  {isSavingOrder ? 'Saving…' : 'Save order'}
-                </button>
-              </>
-            ) : (
-              <button
-                onClick={() => setIsReorderMode(true)}
-                className="px-4 py-2 rounded-lg border border-border bg-card text-foreground font-medium hover:bg-muted transition-colors"
-              >
-                Edit order
-              </button>
-            )}
-          </div>
+          <ProjectsReorderBar
+            isReorderMode={isReorderMode}
+            isSavingOrder={isSavingOrder}
+            onEnterReorderMode={() => setIsReorderMode(true)}
+            onCancelReorder={handleCancelReorder}
+            onSaveOrder={handleSaveOrder}
+          />
         )}
 
         {loading && (
@@ -740,7 +428,7 @@ export default function ProjectsPage() {
               {selectedTag === TAG_NONE
                 ? 'No projects without tags.'
                 : selectedTag
-                  ? `No projects with tag “${selectedTag}”.`
+                  ? `No projects with tag "${selectedTag}".`
                   : selectedYear
                     ? `No projects from ${selectedYear}.`
                     : 'No projects yet.'}
@@ -817,227 +505,28 @@ export default function ProjectsPage() {
           </>
         )}
 
-        {/* Gallery modal: thumbnails for the project; selecting one opens the photo modal at that image */}
-        {galleryModalProject && (() => {
-          const imageUrls =
-            galleryModalProject.images?.length
-              ? galleryModalProject.images.map((i) => i.imageUrl)
-              : galleryModalProject.imageUrl
-                ? [galleryModalProject.imageUrl]
-                : [];
-          if (imageUrls.length === 0) return null;
-          const openPhotoAt = (index: number) => openPhotoModal(galleryModalProject, index);
-          return (
-            <div
-              role="dialog"
-              aria-modal="true"
-              aria-label={`Gallery: ${galleryModalProject.title}`}
-              className="fixed inset-0 z-50 flex flex-col bg-background p-0 sm:p-4 sm:py-6"
-              onClick={() => closeGalleryModal()}
-            >
-              <div className="flex items-center justify-between shrink-0 px-4 py-3 sm:px-0 border-b border-border sm:border-0">
-                <h2 className="text-lg font-semibold text-foreground truncate pr-2">{galleryModalProject.title}</h2>
-                <button
-                  type="button"
-                  onClick={(e) => { e.stopPropagation(); closeGalleryModal(); }}
-                  className="min-h-[44px] min-w-[44px] flex items-center justify-center rounded-full bg-muted text-foreground hover:bg-muted/80 focus:outline-none focus-visible:ring-2 focus-visible:ring-ring"
-                  aria-label="Close gallery"
-                >
-                  <svg className="w-6 h-6" fill="none" stroke="currentColor" viewBox="0 0 24 24" aria-hidden>
-                    <path strokeLinecap="round" strokeLinejoin="round" strokeWidth={2} d="M6 18L18 6M6 6l12 12" />
-                  </svg>
-                </button>
-              </div>
-              <div
-                className="flex-1 overflow-auto p-4 flex flex-wrap items-start gap-3 sm:gap-4"
-                onClick={(e) => e.stopPropagation()}
-              >
-                {imageUrls.map((url, i) => (
-                  <div
-                    key={`${url}-${i}`}
-                    className="w-[calc((100%-0.75rem)/2)] sm:w-52 md:w-64 shrink-0"
-                  >
-                    <div className="relative w-full" style={{ paddingBottom: '100%' }}>
-                      <button
-                        type="button"
-                        onClick={() => openPhotoAt(i)}
-                        className="absolute inset-0 rounded-lg overflow-hidden bg-muted focus:outline-none focus-visible:ring-2 focus-visible:ring-ring focus-visible:ring-offset-2 w-full h-full"
-                        aria-label={`View image ${i + 1} of ${imageUrls.length}`}
-                      >
-                        <Image
-                          src={url}
-                          alt={`${galleryModalProject.title} — image ${i + 1}`}
-                          fill
-                          className="object-cover"
-                          sizes="(max-width: 640px) 50vw, (max-width: 768px) 33vw, 25vw"
-                        />
-                      </button>
-                    </div>
-                  </div>
-                ))}
-              </div>
-            </div>
-          );
-        })()}
+        {galleryModalProject && (
+          <GalleryModal
+            project={galleryModalProject}
+            onClose={closeGalleryModal}
+            onOpenPhoto={(index) => openPhotoModal(galleryModalProject, index)}
+          />
+        )}
 
-        {/* Photo lightbox modal with carousel — only used when not in reorder mode (ProjectCard only) */}
-        {photoModalProject && (() => {
-          const imageUrls =
-            photoModalProject.images?.length
-              ? photoModalProject.images.map((i) => i.imageUrl)
-              : photoModalProject.imageUrl
-                ? [photoModalProject.imageUrl]
-                : [];
-          if (imageUrls.length === 0) return null;
-          const currentUrl = imageUrls[photoModalIndex] ?? imageUrls[0];
-          const canGoPrev = imageUrls.length > 1;
-          const canGoNext = imageUrls.length > 1;
-          const goPrev = () => {
-            setPhotoModalIndex((i) => (i - 1 + imageUrls.length) % imageUrls.length);
-            setPhotoModalDragOffset(0);
-          };
-          const goNext = () => {
-            setPhotoModalIndex((i) => (i + 1) % imageUrls.length);
-            setPhotoModalDragOffset(0);
-          };
-          const handleTouchStart = (e: React.TouchEvent) => {
-            if (imageUrls.length <= 1) return;
-            touchStartX.current = e.touches[0].clientX;
-            touchStartOffset.current = photoModalDragOffset;
-            setPhotoModalIsDragging(true);
-          };
-          const handleTouchMove = (e: React.TouchEvent) => {
-            if (touchStartX.current == null || imageUrls.length <= 1) return;
-            const clientX = e.touches[0].clientX;
-            let rawOffset = touchStartOffset.current + (clientX - touchStartX.current);
-            if (photoModalIndex <= 0 && rawOffset > 0) rawOffset = rawOffset * 0.3;
-            if (photoModalIndex >= imageUrls.length - 1 && rawOffset < 0) rawOffset = rawOffset * 0.3;
-            setPhotoModalDragOffset(rawOffset);
-          };
-          const handleTouchEnd = (e: React.TouchEvent) => {
-            if (touchStartX.current == null || imageUrls.length <= 1) {
-              setPhotoModalIsDragging(false);
-              return;
-            }
-            const dx = e.changedTouches[0].clientX - touchStartX.current;
-            const width = carouselContainerRef.current?.offsetWidth ?? 300;
-            const threshold = Math.min(width * 0.25, 80);
-            if (dx > threshold && photoModalIndex > 0) goPrev();
-            else if (dx < -threshold && photoModalIndex < imageUrls.length - 1) goNext();
-            else setPhotoModalDragOffset(0);
-            touchStartX.current = null;
-            setPhotoModalIsDragging(false);
-          };
-          return (
-            <div
-              role="dialog"
-              aria-modal="true"
-              aria-label={`Viewing ${photoModalProject.title}, image ${photoModalIndex + 1} of ${imageUrls.length}`}
-              className="fixed inset-0 z-50 flex items-center justify-center p-0 sm:p-4 bg-black/90 touch-manipulation"
-              onClick={() => closePhotoModal()}
-            >
-              <button
-                type="button"
-                onClick={(e) => { e.stopPropagation(); closePhotoModal(); }}
-                className="absolute top-4 right-4 z-10 min-h-[44px] min-w-[44px] flex items-center justify-center rounded-full bg-white/15 text-white hover:bg-white/25 focus:outline-none focus-visible:ring-2 focus-visible:ring-white focus-visible:ring-offset-2 focus-visible:ring-offset-black"
-                aria-label="Close"
-              >
-                <svg className="w-6 h-6" fill="none" stroke="currentColor" viewBox="0 0 24 24" aria-hidden>
-                  <path strokeLinecap="round" strokeLinejoin="round" strokeWidth={2} d="M6 18L18 6M6 6l12 12" />
-                </svg>
-              </button>
-
-              {canGoPrev && (
-                <button
-                  type="button"
-                  onClick={(e) => { e.stopPropagation(); goPrev(); }}
-                  className="absolute left-2 sm:left-4 top-1/2 -translate-y-1/2 z-10 min-h-[44px] min-w-[44px] hidden sm:flex items-center justify-center rounded-full bg-white/15 text-white hover:bg-white/25 focus:outline-none focus-visible:ring-2 focus-visible:ring-white focus-visible:ring-offset-2 focus-visible:ring-offset-black"
-                  aria-label="Previous image"
-                >
-                  <svg className="w-8 h-8" fill="none" stroke="currentColor" viewBox="0 0 24 24" aria-hidden>
-                    <path strokeLinecap="round" strokeLinejoin="round" strokeWidth={2} d="M15 19l-7-7 7-7" />
-                  </svg>
-                </button>
-              )}
-              {canGoNext && (
-                <button
-                  type="button"
-                  onClick={(e) => { e.stopPropagation(); goNext(); }}
-                  className="absolute right-2 sm:right-4 top-1/2 -translate-y-1/2 z-10 min-h-[44px] min-w-[44px] hidden sm:flex items-center justify-center rounded-full bg-white/15 text-white hover:bg-white/25 focus:outline-none focus-visible:ring-2 focus-visible:ring-white focus-visible:ring-offset-2 focus-visible:ring-offset-black"
-                  aria-label="Next image"
-                >
-                  <svg className="w-8 h-8" fill="none" stroke="currentColor" viewBox="0 0 24 24" aria-hidden>
-                    <path strokeLinecap="round" strokeLinejoin="round" strokeWidth={2} d="M9 5l7 7-7 7" />
-                  </svg>
-                </button>
-              )}
-
-              <div
-                ref={carouselContainerRef}
-                className="relative w-full h-full sm:max-w-4xl sm:h-[85vh] flex items-center overflow-hidden touch-pan-y"
-                onClick={(e) => e.stopPropagation()}
-                onTouchStart={handleTouchStart}
-                onTouchMove={handleTouchMove}
-                onTouchEnd={handleTouchEnd}
-              >
-                {imageUrls.length > 1 ? (
-                  <div
-                    className="flex h-full shrink-0"
-                    style={{
-                      width: `${imageUrls.length * 100}%`,
-                      transform: `translateX(calc(-${(100 * photoModalIndex) / imageUrls.length}% + ${photoModalDragOffset}px))`,
-                      transition: photoModalIsDragging ? 'none' : 'transform 0.25s ease-out',
-                    }}
-                  >
-                    {imageUrls.map((url, i) => (
-                      <div
-                        key={`${url}-${i}`}
-                        className="relative h-full shrink-0 flex items-center justify-center"
-                        style={{ width: `${100 / imageUrls.length}%` }}
-                      >
-                        <Image
-                          src={url}
-                          alt={`${photoModalProject.title} — image ${i + 1} of ${imageUrls.length}`}
-                          fill
-                          className="object-contain sm:rounded-lg pointer-events-none"
-                          sizes="(max-width: 640px) 100vw, 896px"
-                          priority={i <= 1}
-                        />
-                      </div>
-                    ))}
-                  </div>
-                ) : (
-                  <div className="relative w-full h-full flex items-center justify-center">
-                    <Image
-                      src={currentUrl}
-                      alt={`${photoModalProject.title} — image 1 of 1`}
-                      fill
-                      className="object-contain sm:rounded-lg"
-                      sizes="(max-width: 640px) 100vw, 896px"
-                      priority
-                    />
-                  </div>
-                )}
-              </div>
-
-              {imageUrls.length > 1 && (
-                <div
-                  className="absolute bottom-4 left-1/2 -translate-x-1/2 z-10 flex items-center gap-1.5 rounded-full bg-black/50 backdrop-blur-sm px-2 py-1.5"
-                  aria-hidden
-                >
-                  {imageUrls.map((_, i) => (
-                    <span
-                      key={i}
-                      className={`rounded-full transition-all ${
-                        i === photoModalIndex ? 'h-1.5 w-1.5 bg-white' : 'h-1 w-1 bg-white/50'
-                      }`}
-                    />
-                  ))}
-                </div>
-              )}
-            </div>
-          );
-        })()}
+        {photoModalProject && photoModalImageUrls.length > 0 && (
+          <PhotoModal
+            project={photoModalProject}
+            imageUrls={photoModalImageUrls}
+            index={photoModalIndex}
+            onIndexChange={setPhotoModalIndex}
+            dragOffset={photoModalDragOffset}
+            onDragOffsetChange={setPhotoModalDragOffset}
+            isDragging={photoModalIsDragging}
+            onIsDraggingChange={setPhotoModalIsDragging}
+            containerRef={carouselContainerRef}
+            onClose={closePhotoModal}
+          />
+        )}
       </div>
     </div>
   );
